Bot_Email/requirements.txt
requests
python-dotenv
google-generativeai
python-telegram-bot
